attach >>| control time_to_cursor fix
git-svn-id: https://svn.xiph.org/trunk/postfish@5479 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ input.h
@@ -8,4 +8,4 @@ extern time_linkage *input_read(void);
 extern int input_load(int n,char *list[]);
 extern int pull_input_feedback(double *peak,double *rms,off_t *cursor,int *n);
 extern void input_reset(void);
-extern int input_time_seek_rel(double s);
+extern off_t input_time_seek_rel(double s);
